refactor sanity queries out of instructor.ts
